Na tela do avaliador, corrige o avanço automático das avaliações ao finalizar...

See merge request mapa-cultural/mapa-cultural!42

+++ src/protected/application/themes/BaseV1/layouts/parts/panel-event.php
@@ -1,3 +1,9 @@
+<?php
+use MapasCulturais\Entities\Event;
+use MapasCulturais\i;
+
+$event = $app->repo('Event')->find($entity->id);
+?>
 <article class="objeto clearfix">
     <h1><a href="<?php echo $entity->singleUrl; ?>"><?php echo $entity->name; ?></a></h1>
     <div class="objeto-meta">
@@ -13,12 +19,15 @@
         <?php endif; ?>
 	</div>
     <div class="entity-actions">
+        <?php $this->applyTemplateHook('entity-actions','begin', [ $entity ]); ?>
         <a class="btn btn-small btn-primary" href="<?php echo $entity->editUrl; ?>"><?php \MapasCulturais\i::_e("editar");?></a>
         <?php if(!isset($only_edit_button)): ?>
             <?php if($entity->status === \MapasCulturais\Entities\Event::STATUS_ENABLED): ?>
 
                 <a class="btn btn-small btn-warning" href="<?php echo $entity->unpublishUrl; ?>"><?php \MapasCulturais\i::_e("tornar rascunho");?></a>
-                <a class="btn btn-small btn-danger" href="<?php echo $entity->deleteUrl; ?>"><?php \MapasCulturais\i::_e("excluir");?></a>
+                <?php if($event->canUser('remove')): ?>
+                    <a class="btn btn-small btn-danger" href="<?php echo $entity->deleteUrl; ?>"><?php \MapasCulturais\i::_e("excluir");?></a>
+                <?php endif; ?>
                 <a class="btn btn-small btn-success" href="<?php echo $entity->archiveUrl; ?>"><?php \MapasCulturais\i::_e("arquivar");?></a>
 
             <?php elseif ($entity->status === \MapasCulturais\Entities\Event::STATUS_DRAFT): ?>
@@ -30,10 +39,13 @@
             <?php else: ?>
                 <a class="btn btn-small btn-success" href="<?php echo $entity->undeleteUrl; ?>"><?php \MapasCulturais\i::_e("recuperar");?></a>
                 <?php if($entity->permissionTo->destroy): ?>
-                    <a class="btn btn-small btn-danger" href="<?php echo $entity->destroyUrl; ?>"><?php \MapasCulturais\i::_e("excluir definitivamente");?></a>
+                    <?php if($event->canUser('destroy')): ?>
+                        <a class="btn btn-small btn-danger" href="<?php echo $entity->destroyUrl; ?>"><?php \MapasCulturais\i::_e("excluir definitivamente");?></a>
+                    <?php endif; ?>
                 <?php endif; ?>
             <?php endif; ?>
 
         <?php endif; ?>
+        <?php $this->applyTemplateHook('entity-actions','end', [ $entity ]); ?>
     </div>
 </article>
