+++ src/protected/application/themes/BaseV1/layouts/parts/panel-subsite.php
@@ -20,10 +20,13 @@
 		<div><span class="label"><?php \MapasCulturais\i::_e("Data de Criação:");?></span> <?php echo $entity->createTimestamp->format('d/m/Y H:i:s');?></div>
 	</div>
     <div class="entity-actions">
+        <?php $this->applyTemplateHook('entity-actions','begin', [ $entity ]); ?>
         <a class="btn btn-small btn-primary" href="<?php echo $entity->editUrl; ?>"><?php \MapasCulturais\i::_e("editar");?></a>
 
         <?php if($entity->status === \MapasCulturais\Entities\Subsite::STATUS_ENABLED): ?>
-            <a class="btn btn-small btn-danger" href="<?php echo $entity->deleteUrl; ?>"><?php \MapasCulturais\i::_e("excluir");?></a>
+            <?php if($entity->canUser('remove')): ?>
+                <a class="btn btn-small btn-danger" href="<?php echo $entity->deleteUrl; ?>"><?php \MapasCulturais\i::_e("excluir");?></a>
+            <?php endif; ?>
             <a class="btn btn-small btn-success" href="<?php echo $entity->archiveUrl; ?>"><?php \MapasCulturais\i::_e("arquivar");?></a>
 
         <?php elseif ($entity->status === \MapasCulturais\Entities\Subsite::STATUS_DRAFT): ?>
@@ -39,5 +42,6 @@
                 <a class="btn btn-small btn-danger" href="<?php echo $entity->destroyUrl; ?>"><?php \MapasCulturais\i::_e("excluir definitivamente");?></a>
             <?php endif; ?>
         <?php endif; ?>
+        <?php $this->applyTemplateHook('entity-actions','end', [ $entity ]); ?>
     </div>
 </article>
